Refactor with isort to organize imports
