+++ calculator/facade/history.csv
@@ -1,0 +1,2 @@
+Operation,Operand1,Operand2,Result
+ADD,79,25,104
