+++ calculator/factory/operation_factory.py
@@ -2,7 +2,6 @@ from calculator.commands.add import AddCommand
 from calculator.commands.divide import DivideCommand
 from calculator.commands.multiply import MultiplyCommand
 from calculator.commands.subtract import SubtractCommand
-from calculator.strategy.average import AverageOperation
 from calculator.strategy.power import PowerOperation
 
 
@@ -17,7 +16,6 @@ class OperationFactory:
             "MULTIPLY": MultiplyCommand,
             "DIVIDE": DivideCommand,
             "POWER": PowerOperation,
-            "AVERAGE": AverageOperation,
         }
         command = command.upper()
         if command not in commands:
