--- calculator/strategy/average.py
@@ -1,7 +1,0 @@
-from .custom_operation import CustomOperation
-
-
-class AverageOperation(CustomOperation):
-    """Raises a number to the power of another."""
-
-    print("Average operation")
